#12 and #28 : split up top bar and side bar, added validators for createWorkspaceForm, cleaned up and abstracted code
