--- src/app/components/TopAndSideBar/index.tsx
@@ -1,384 +1,0 @@
-import * as React from 'react'
-import { Link } from 'react-router-dom'
-import { observer, inject } from 'mobx-react';
-import { AUTH_STORE } from 'app/constants'
-import { AuthStore } from 'app/stores';
-// import * as style from './style.scss'
-import clsx from 'clsx';
-import { makeStyles, fade } from '@material-ui/core/styles';
-import Drawer from '@material-ui/core/Drawer';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import List from '@material-ui/core/List';
-import Divider from '@material-ui/core/Divider';
-import IconButton from '@material-ui/core/IconButton';
-import MenuIcon from '@material-ui/icons/Menu';
-import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemIcon from '@material-ui/core/ListItemIcon';
-import InputBase from '@material-ui/core/InputBase';
-
-// form dialogue 
-import Button from '@material-ui/core/Button';
-import TextField from '@material-ui/core/TextField';
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import InputAdornment from '@material-ui/core/InputAdornment';
-
-
-// icons
-
-import DashboardIcon from '@material-ui/icons/Dashboard';
-import DescriptionIcon from '@material-ui/icons/Description';
-import FormatListBulletedIcon from '@material-ui/icons/FormatListBulleted';
-import CodeIcon from '@material-ui/icons/Code';
-import ListItemText from '@material-ui/core/ListItemText';
-import MenuItem from '@material-ui/core/MenuItem';
-import Menu from '@material-ui/core/Menu';
-import AccountCircle from '@material-ui/icons/AccountCircle';
-import SearchIcon from '@material-ui/icons/Search';
-import AddCircleOutlineIcon from '@material-ui/icons/AddCircleOutline';
-import MemoryIcon from '@material-ui/icons/Memory';
-import MoneyIcon from '@material-ui/icons/Money';
-import StorageIcon from '@material-ui/icons/Storage';
-
-/**
- * Code Sourced from: https://github.com/mui-org/material-ui/tree/master/docs/src/pages/getting-started/templates/dashboard
- * https://material-ui.com/getting-started/templates/dashboard/
- */
-
-
-
-const drawerWidth = 240;
-
-
-const useStyles = makeStyles((theme) => ({
-    toolbar: {
-        paddingRight: 24, // keep right padding when drawer closed
-    },
-    toolbarIcon: {
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'flex-end',
-        padding: '0 8px',
-        ...theme.mixins.toolbar,
-    },
-    appBar: {
-        zIndex: theme.zIndex.drawer + 1,
-        transition: theme.transitions.create(['width', 'margin'], {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.leavingScreen,
-        }),
-    },
-    appBarShift: {
-        marginLeft: drawerWidth,
-        width: `calc(100% - ${drawerWidth}px)`,
-        transition: theme.transitions.create(['width', 'margin'], {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.enteringScreen,
-        }),
-    },
-    menuButton: {
-        marginRight: 36,
-    },
-    menuButtonHidden: {
-        display: 'none',
-    },
-    title: {
-        flexGrow: 1,
-    },
-    drawerPaper: {
-        position: 'relative',
-        whiteSpace: 'nowrap',
-        width: drawerWidth,
-        transition: theme.transitions.create('width', {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.enteringScreen,
-        }),
-    },
-    drawerPaperClose: {
-        overflowX: 'hidden',
-        transition: theme.transitions.create('width', {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.leavingScreen,
-        }),
-        width: theme.spacing(7),
-        [theme.breakpoints.up('sm')]: {
-            width: theme.spacing(9),
-        },
-    },
-    paper: {
-        padding: theme.spacing(2),
-        display: 'flex',
-        overflow: 'auto',
-        flexDirection: 'column',
-    },
-    search: {
-        position: 'relative',
-        borderRadius: theme.shape.borderRadius,
-        backgroundColor: fade(theme.palette.common.white, 0.15),
-        '&:hover': {
-          backgroundColor: fade(theme.palette.common.white, 0.25),
-        },
-        marginLeft: 0,
-        width: '100%',
-        [theme.breakpoints.up('sm')]: {
-          marginLeft: theme.spacing(1),
-          width: 'auto',
-        },
-    },
-    searchIcon: {
-        padding: theme.spacing(0, 2),
-        height: '100%',
-        position: 'absolute',
-        pointerEvents: 'none',
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-    },
-    inputRoot: {
-        color: 'inherit',
-    },
-    inputInput: {
-        padding: theme.spacing(1, 1, 1, 0),
-        // vertical padding + font size from searchIcon
-        paddingLeft: `calc(1em + ${theme.spacing(4)}px)`,
-        transition: theme.transitions.create('width'),
-        width: '100%',
-        [theme.breakpoints.up('sm')]: {
-          width: '12ch',
-          '&:focus': {
-            width: '20ch',
-          },
-        },
-    },
-    accountIcon: {
-        position: 'absolute',
-        right: '0px',
-    },
-    margin: {
-        margin: theme.spacing(1, 0, 1, 0),
-    },
-}));
-
-
-export const TopAndSideBar = inject(AUTH_STORE)(observer((props) => {
-    const store = props[AUTH_STORE] as AuthStore
-    const classes = useStyles();
-
-    //account icon handlers
-    const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
-    const openAccount = Boolean(anchorEl);
-    const handleAccountMenu = (event: React.MouseEvent<HTMLElement>) => {
-        setAnchorEl(event.currentTarget);
-    };
-    const handleIconClose = () => {
-        setAnchorEl(null);
-    };
-    const handleSignOut = async () => {
-        await store.logout();
-        setAnchorEl(null);
-    }
-
-    //drawer handlers
-    const [open, setOpen] = React.useState(false);
-    const handleDrawerOpen = () => {
-        setOpen(true);
-    };
-    const handleDrawerClose = () => {
-        setOpen(false);
-    };
-
-    //form dialog
-    const [openForm, setOpenForm] = React.useState(false);
-
-    const handleClickOpen = () => {
-        setOpenForm(true);
-    };
-
-    const handleClose = () => {
-        setOpenForm(false);
-    };
-
-    return <> 
-        <AppBar position="absolute" className={clsx(classes.appBar, open && classes.appBarShift)}>
-            <Toolbar className={classes.toolbar}>
-                <IconButton
-                    edge="start"
-                    color="inherit"
-                    aria-label="open drawer"
-                    onClick={handleDrawerOpen}
-                    className={clsx(classes.menuButton, open && classes.menuButtonHidden)}
-                >
-                <MenuIcon />
-                </IconButton>
-                {props.children}
-                <div className={classes.search}>
-                    <div className={classes.searchIcon}>
-                        <SearchIcon />
-                    </div>
-                    <InputBase
-                    placeholder="Search…"
-                    classes={{
-                        root: classes.inputRoot,
-                        input: classes.inputInput,
-                    }}
-                    inputProps={{ 'aria-label': 'search' }}
-                    />
-                </div>
-                <div className={classes.accountIcon}>
-                    <IconButton
-                        aria-label="account of current user"
-                        aria-controls="menu-appbar"
-                        aria-haspopup="true"
-                        onClick={handleAccountMenu}
-                        color="inherit"
-                    >
-                        <AccountCircle />
-                    </IconButton>
-                    <Menu
-                        id="menu-appbar"
-                        anchorEl={anchorEl}
-                        anchorOrigin={{
-                            vertical: 'top',
-                            horizontal: 'right',
-                        }}
-                        keepMounted
-                        transformOrigin={{
-                            vertical: 'top',
-                            horizontal: 'right',
-                        }}
-                        open={openAccount}
-                        onClose={handleIconClose}
-                    >
-                        <MenuItem onClick={handleIconClose}>My Profile</MenuItem>
-                        <MenuItem onClick={handleSignOut}>Sign Out</MenuItem>
-                    </Menu>
-                </div>
-            </Toolbar>
-        </AppBar>
-        <Drawer
-            variant="permanent"
-            classes={{
-                paper: clsx(classes.drawerPaper, !open && classes.drawerPaperClose),
-            }}
-            open={open}
-        >
-            <div className={classes.toolbarIcon}>
-                <IconButton onClick={handleDrawerClose}>
-                    <ChevronLeftIcon />
-                </IconButton>
-            </div>
-            <Divider />
-            <List>
-                <ListItem button component={Link} to="">
-                    <ListItemIcon>
-                        <DashboardIcon />
-                    </ListItemIcon>
-                    <ListItemText primary="Deployments" />
-                </ListItem>
-                <ListItem button component={Link} to="">
-                    <ListItemIcon>
-                        <CodeIcon />
-                    </ListItemIcon>
-                    <ListItemText primary="Code" />
-                </ListItem>
-                <ListItem button component={Link} to="">
-                    <ListItemIcon>
-                        <DescriptionIcon />
-                    </ListItemIcon>
-                    <ListItemText primary="Wiki" />
-                </ListItem>
-                <ListItem button component={Link} to="">
-                    <ListItemIcon>
-                        <FormatListBulletedIcon />
-                    </ListItemIcon>
-                    <ListItemText primary="Tasks" />
-                </ListItem>
-                <ListItem button onClick={handleClickOpen}>  
-                    <ListItemIcon>
-                        <AddCircleOutlineIcon />
-                    </ListItemIcon>
-                    <ListItemText primary="Create Workspace" />    
-                </ListItem>
-            </List>
-        </Drawer>
-        <Dialog open={openForm} onClose={handleClose} aria-labelledby="form-dialog-title">
-        <DialogTitle id="form-dialog-title">Create Workspace</DialogTitle>
-        <DialogContent>
-          <DialogContentText>
-            Please fill out the form below and press 'Create' to create a workspace.
-          </DialogContentText>
-          <TextField
-            autoFocus
-            className={classes.margin}
-            variant="outlined"
-            margin="dense"
-            id="name"
-            label="Workspace Name"
-            type="text"
-            fullWidth
-          />
-          <TextField
-            className={classes.margin}
-            margin="dense"
-            variant="outlined"
-            id="cpu"
-            label="CPU"
-            type="number"
-            InputProps={{
-                startAdornment: (
-                    <InputAdornment position="start">
-                        <MemoryIcon />
-                    </InputAdornment>
-                ),
-            }}
-            fullWidth
-            />
-            <TextField
-            className={classes.margin}
-            margin="dense"
-            variant="outlined"
-            id="ram"
-            label="RAM"
-            type="number"
-            InputProps={{
-                startAdornment: (
-                    <InputAdornment position="start">
-                        <MoneyIcon />
-                    </InputAdornment>
-                ),
-            }}
-            fullWidth
-            />
-            <TextField
-            className={classes.margin}
-            margin="dense"
-            variant="outlined"
-            id="storage"
-            label="Storage"
-            type="number"
-            InputProps={{
-                startAdornment: (
-                    <InputAdornment position="start">
-                        <StorageIcon />
-                    </InputAdornment>
-                ),
-            }}
-            fullWidth
-            />
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleClose} color="primary">
-            Cancel
-          </Button>
-          <Button onClick={handleClose} color="primary">
-            Create
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </>
-}))
