+++ src/app/forms/Workspace/index.ts
@@ -1,0 +1,1 @@
+export * from './CreateWorkspace'
