+++ src/app/forms/index.ts
@@ -1,2 +1,3 @@
 export * from './Auth'
 export * from './Base'
+export * from './Workspace'
